adding accountId queries to support multiple accounts

+++ Alice/SignedPreKey.cpp
@@ -7,11 +7,12 @@
 using namespace std;
 using namespace sqlite;
 
-CriptextDB::SignedPreKey CriptextDB::getSignedPreKey(database db, short int id) {
+CriptextDB::SignedPreKey CriptextDB::getSignedPreKey(database db, int accountId, short int id) {
   string mySignedPreKey = "";
   int myLen;
-  db << "Select * from signedprekeyrecord where signedPreKeyId == ?;"
+  db << "Select * from signedprekeyrecord where signedPreKeyId == and accountId == ?;"
      << id
+     << accountId
      >> [&] (int preKeyId, string record, int recordLength) {
         mySignedPreKey = record;
         myLen = (size_t)recordLength;
@@ -28,12 +29,31 @@ CriptextDB::SignedPreKey CriptextDB::getSignedPreKey(database db, short int id) 
   return signedPreKey;
 }
 
-bool CriptextDB::createSignedPreKey(database db, short int id, char* keyRecord, size_t len) {
+bool CriptextDB::createSignedPreKey(database db, int accountId, short int id, char* keyRecord, size_t len) {
   try {
-    db << "insert into signedprekeyrecord (signedPreKeyId, record, recordLength) values (?,?,?);"
-     << id
-     << keyRecord
-     << static_cast<int>(len);
+    bool hasRow = false;
+    db << "begin;";
+    db << "Select signedPreKeyId from signedprekeyrecord where signedPreKeyId == ? and accountId == ?;"
+      << id
+      << accountId
+      >> [&](string signedPreKeyId) {
+      hasRow = true;
+    };
+    if (hasRow) {
+      db << "update signedprekeyrecord set record = ?, recordLength = ? where signedPreKeyId == ? and accountId == ?;"
+        << keyRecord
+        << static_cast<int>(len)
+        << id
+        << accountId;
+    }
+    else {
+      db << "insert into signedprekeyrecord (signedPreKeyId, record, recordLength, accountId) values (?,?,?,?);"
+        << id
+        << keyRecord
+        << static_cast<int>(len)
+        << accountId;
+    }
+    db << "commit;";
     return true;
   } catch (exception& e) {
     std::cout << e.what() << std::endl;
@@ -41,10 +61,11 @@ bool CriptextDB::createSignedPreKey(database db, short int id, char* keyRecord, 
   }
 }
 
-bool CriptextDB::deleteSignedPreKey(database db, short int id) {
+bool CriptextDB::deleteSignedPreKey(database db, int accountId, short int id) {
   try {
-    db << "delete from signedPrekeyrecord where signedPreKeyId == ?;"
-     << id;
+    db << "delete from signedPrekeyrecord where signedPreKeyId == ? and accountId == ?;"
+      << id
+      << accountId;
     return true;
   } catch (exception& e) {
     std::cout << e.what() << std::endl;
